requirements.txt
pinecone-client
langchain~=0.0.198
streamlit~=1.23.1
openai~=0.27.8
python-dotenv~=1.0.0